create client up front
